+++ Change Log.md
@@ -1,3 +1,27 @@
+v 0.1.4
+- Added chapters feed generation for translating PeerTube chapters to Podcastion 2.0 chapters
+- Cleaned up RSS setting dialog
+
+v 0.1.3
+- fixed hive interaction
+
+v 0.1.2
+- fixed new podcast guid generation
+- fixed guid interactions with lightning plugin
+
+v 0.1.1
+- updated new lightning address attribute name to keysend
+
+v 0.1.0
+- added itunes:type serial to audiobook feeds
+- added embedded peertube alternate enclosure
+- added youtube source ID to plugin video fields
+- added embedded youtube alternate enclosure
+- fixed mime type for enclosure to application/x-mpegURL
+
+v 0.0.9
+- another attempt at fixing axios import on some instances
+
 v 0.0.8
 - ?fixed bug with missing dependencies
 - fixed audio stream type if no video present in enclosure file